src/discord_bot/requirements.txt
requests
pynacl
supabase==2.20.0